Replace Glib::ThreadPool with custom work queue
This is a pretty big refactor of the ImageWorker stuff. The main benefit
is that the work queue can be cleared, so ImageList can cancel all of
its pending thumbnails when opening a new folder. Hopefully there's more
flexibility for additions later on, too.

Change the comment style for class/function descriptions (will make
other files consistent in a followup commit). Also turn on more GCC
warnings.