Use Autotools build system
This replaces the plain Makefile that was being used. Checks are made
for a minimum gtkmm version and C++11 support now.

Some additional runtime checks are needed to find the data directory
when the executable is either installed or running from the build
directory. This is done in utils.cpp's get_data_dir() and isn't a
perfect solution, but works okay for developing without having to
install to test changes.