Use per-task slot in ImageWorker
Instead of having a class signal that you connect to once, require a
slot for every call to load(). This allows variables unique to each task
to be bound to the slot, reducing coupling between ImageWorker and its
callers. (Namely, the dependency on Gtk::TreeIter in image_worker.h is
removed.)